New feature: Adds dynamic on-page navigation
Implements a new "On This Page" component that provides
dynamic navigation based on headings within the main content.
It utilizes an IntersectionObserver to highlight the currently
visible heading.

Updates the Layout component to only include h2 elements in the
table of contents and uses the 'location.pathname' to update
the table of contents for each page.