New feature: Adds donation banner
Adds a donation banner to the site to encourage users to support the project.

The banner appears at the top of the page and includes a link to Ko-fi.
It can be dismissed by the user, and will reappear after one week.

The banner has different layouts for desktop and mobile/medium screens to
ensure a good user experience on all devices.