Maintenance: Migrates to Vite React SSG
Migrates the documentation site to Vite React SSG for improved performance and static site generation.

This change replaces React Router with Vite React SSG's routing mechanism and updates the Layout component to use Outlet for rendering child routes. It also updates the build and dev scripts in package.json and updates the OnThisPage component to work with client side rendering.